refactor: centralize ui models to one package
