+++ features/post/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/post/presentation/models/AuthorUi.kt
@@ -1,4 +1,4 @@
-package com.parksupark.soomjae.features.post.presentation.post.models
+package com.parksupark.soomjae.features.post.presentation.models
 
 import com.parksupark.soomjae.features.post.domain.models.Member
