+++ features/post/presentation/src/commonMain/kotlin/com/parksupark/soomjae/features/post/presentation/post/tabs/community/components/CommunityPostCard.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
 import com.parksupark.soomjae.core.presentation.designsystem.theme.SoomjaeTheme
 import com.parksupark.soomjae.core.presentation.ui.utils.imageRequest
-import com.parksupark.soomjae.features.post.presentation.post.models.AuthorUi
-import com.parksupark.soomjae.features.post.presentation.post.models.CommunityPostUi
+import com.parksupark.soomjae.features.post.presentation.models.AuthorUi
+import com.parksupark.soomjae.features.post.presentation.models.CommunityPostUi
 
 @Composable
 internal fun CommunityPostCard(
